+++ src/html/oninstall.html
@@ -1,0 +1,84 @@
+<!doctype html>
+<html lang="en" class="h-100">
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>Simple Extension Permissions</title>
+    <link rel="icon" href="../media/favicon.ico" sizes="any">
+    <link rel="stylesheet" type="text/css" href="../dist/bootstrap/bootstrap.min.css">
+    <link rel="stylesheet" type="text/css" href="../dist/fontawesome/css/all.min.css">
+    <link rel="stylesheet" type="text/css" href="../css/main.css">
+    <script type="text/javascript" src="../js/theme.js"></script>
+</head>
+<body class="h-100">
+
+<div class="container-fluid p-3 h-100">
+    <div class="d-flex align-items-center h-100">
+        <div class="col-xl-6 col-md-8 col-12 m-auto">
+            <div class="card p-3 mb-3">
+                <div class="d-flex justify-content-center align-items-center mb-3">
+                    <img src="../media/logo48.png" class="me-2" height="48" width="48"
+                         alt="Zipline Extension" title="Zipline Extension">
+                    <h1>Zipline Extension</h1>
+                </div>
+
+                <div>
+                    <h2>Automatic Setup</h2>
+                    <ol>
+                        <li>Navigate to your Zipline Dashboard.</li>
+                        <li>
+                            Activate the <a class="open-popup" href="#0">Popup</a>.
+                            <ul>
+                                <li>Click the Toolbar Icon.</li>
+                                <li>Use the Right Click Menu.</li>
+                                <li>Use the Keyboard Shortcut.</li>
+                            </ul>
+                        </li>
+                        <li>Click Authenticate with Current site.</li>
+                    </ol>
+                    <div class="text-center mb-3">
+                    <img alt="Authenticate with Current" class="img-fluid" src="../media/auth.jpg" width="390" height="227">
+                    </div>
+
+                    <h2>Manual Setup</h2>
+                    <ol>
+                        <li>Enter your Zipline URL and Token in the Options.</li>
+                    </ol>
+                    <a class="btn btn-lg btn-outline-info w-100 mb-3" href="../html/options.html">
+                        <i class="fa-solid fa-sliders me-1"></i> Open Options</a>
+
+                    <h2>Support</h2>
+                    <p>
+                        If you find any bugs or need help, please
+                        <a title="Open an Issue" class="fw-bold" target="_blank" rel="noopener"
+                           href="https://github.com/cssnr/zipline-extension/issues">Open an Issue on GitHub</a>.
+                    </p>
+
+                </div>
+
+                <hr>
+
+                <div class="text-center">
+                    <a class="link-body-emphasis text-decoration-none" target="_blank" rel="noopener"
+                       href="https://github.com/cssnr/zipline-extension">View Source</a>
+                    <span class="mx-2">&bull;</span>
+                    <a class="link-body-emphasis text-decoration-none" target="_blank" rel="noopener"
+                       href="https://github.com/cssnr/zipline-extension?tab=readme-ov-file#readme">Get Support</a>
+                    <span class="mx-2">&bull;</span>
+                    <a class="link-body-emphasis text-decoration-none" target="_blank" rel="noopener"
+                       href="https://github.com/cssnr/zipline-android?tab=readme-ov-file#readme">Zipline Android</a>
+                </div>
+            </div> <!-- card -->
+        </div> <!-- col -->
+    </div> <!-- flex -->
+</div> <!-- container -->
+
+<div id="clone" class="d-none">
+</div> <!-- clone -->
+
+<script type="text/javascript" src="../dist/jquery/jquery.min.js"></script>
+<script type="text/javascript" src="../dist/bootstrap/bootstrap.bundle.min.js"></script>
+<script type="module" src="../js/oninstall.js"></script>
+
+</body>
+</html>
