+++ src/js/oninstall.js
@@ -1,0 +1,19 @@
+// JS for oninstall.html
+
+import { linkClick, openPopupPanel } from './exports.js'
+
+document.addEventListener('DOMContentLoaded', domContentLoaded)
+document
+    .querySelectorAll('a[href]')
+    .forEach((el) => el.addEventListener('click', linkClick))
+document
+    .querySelectorAll('.open-popup')
+    .forEach((el) => el.addEventListener('click', openPopupPanel))
+
+/**
+ * DOMContentLoaded
+ * @function domContentLoaded
+ */
+async function domContentLoaded() {
+    console.debug('domContentLoaded')
+}
